+++ file_windows_test.go
@@ -854,6 +854,34 @@ func TestFileWindowsRemoveAll(t *testing.T) {
 	}
 }
 
+func TestFileWindowsRemoveAllRetry(t *testing.T) {
+	for _, perm := range perms {
+		name, err := tempFile(t)
+		if err != nil {
+			t.Fatal(err)
+		}
+		cleanup(t, name)
+
+		err = compat.Chmod(name, perm)
+		checkPerm(t, name, perm, false)
+		if err != nil {
+			t.Fatalf("Chmod(%04o) failed: %v", perm, err)
+		}
+
+		perm = perm777
+		err = compat.Chmod(name, perm)
+		checkPerm(t, name, perm, false)
+		if err != nil {
+			t.Fatalf("Chmod(%04o) failed: %v", perm, err)
+		}
+		err = compat.RemoveAll(name, compat.WithRetrySeconds(2))
+		checkDeleted(t, name, perm, err)
+		if err != nil {
+			t.Fatalf("RemoveAll failed: %v: %v", name, err)
+		}
+	}
+}
+
 func TestFileWindowsWriteFile(t *testing.T) {
 	for _, perm := range perms {
 		name, err := tempName(t)