+++ rename.go
@@ -6,6 +6,6 @@ package compat
 // Rename atomically replaces the destination file or directory with the
 // source. It is guaranteed to either replace the target file entirely, or not
 // change either file.
-func Rename(source, destination string) error {
-	return rename(source, destination)
+func Rename(source, destination string, opts ...Option) error {
+	return rename(source, destination, opts...)
 }