+++ rename_unix.go
@@ -12,4 +12,9 @@ import (
 	"os"
 )
 
-var rename = os.Rename
+// Rename atomically replaces the destination file or directory with the
+// source. It is guaranteed to either replace the target file entirely, or not
+// change either file.
+func Rename(source, destination string, _ ...Option) error {
+	return os.Rename(source, destination)
+}