+++ robustio/robustio_darwin.go
@@ -1,0 +1,29 @@
+// SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
+// SPDX-License-Identifier: MIT
+
+// SPDX-FileCopyrightText: Copyright 2019 The Go Authors.
+// SPDX-License-Identifier: BSD-3
+
+// Source: https://github.com/golang/go/blob/f15cd63e/src/cmd/internal/robustio/robustio_windows.go
+
+// Copyright 2019 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package robustio
+
+import (
+	"errors"
+	"syscall"
+)
+
+const errFileNotFound = syscall.ENOENT
+
+// isEphemeralError returns true if err may be resolved by waiting.
+func isEphemeralError(err error) bool {
+	var errno syscall.Errno
+	if errors.As(err, &errno) {
+		return errno == errFileNotFound
+	}
+	return false
+}